Finishes re-adding Nether Ores exploding and those explosions angering all pigmen in a given radius - also fixes a minor bug in the Tinkers' interactions code
